requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
blinker==1.8.2
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
decorator==5.1.1
exceptiongroup==1.2.2
Flask==3.0.3
google-auth==2.40.1
google-genai==1.14.0
h11==0.16.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.9
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==4.9.4
MarkupSafe==3.0.2
plivo==4.56.2
priority==2.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
python-dotenv==1.0.1
Quart==0.19.8
requests==2.32.3
rsa==4.9.1
six==1.16.0
sniffio==1.3.1
taskgroup==0.0.0a4
tomli==2.0.2
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
Werkzeug==3.1.2
wsproto==1.2.0